moved post answer modal component
